Reorganize files with utils file
