+++ app/assets/src/components/views/PipelineViz/pipeline_viz.scss
